refactor time control config
Previously setting a time control is very confusing as timeout_turn=0 means play as fast as possible rule. This patch removes this rule as this causes problems to the meaning of timeout_turn and timrout_match.
A non-positive timeout_turn or timeout_match will mean "infinite time mode" from now on. And time_left is automatically subtracted at each move even if it's not set from the protocol.

[skip_fishtest]