more robust handling of multilingual text encoding
Multilingual text encoding was not well implemented before, which may cause garbled characters under Windows (Linux would be fine where everything is UTF-8). This patch aims to fix all multilingual text encoding issues with careful treatment.

Currently there are several places where non-ASCII characters can appear:
1. console input/output (for specifying a path or a database text comment);
2. command arguments in argv (for specifying a path);
3. content of a config file (for specifying path/URL to weight or database)
4. dbtotxt functionality that contains text comments

For internal representation of path, we use std::filesystem::path and convert plain string to it with care of converting the encoding. For internal representation of URL and all database texts, we use UTF-8.

We use `GetConsoleCP()` and `GetConsoleOutputCP()` to determine text encoding of std::cin and std::cout for (1) and (2). The console codepage can be explicitly set to UTF-8 by specifying in the startup argument, if other program wants to communicate with rapfi in UTF-8. The `chcp` command on Windows can be also used to change the console codepage.

For (3) and (4), we always assume content read/write in config file or other text file is in UTF-8. Note that some old database file may use local codepage due to early mistaken implementation. We have started to put a meta record in newer database some time ago, to explicitly states we are using UTF-8 as the charset. 

When reading early `.db` files that have no UTF-8 meta record and imported `.lib` files on Windows, we use the "legacy_file_code_page" option in config to determine their encoding (which defaults to 0 for Windows's active code page). Note that "legacy_file_code_page" is only supported on Windows, thus when reading on linux or other system, we always treat them as in UTF-8. We plan to forbid reading of those early `.db` files on other systems in the future.