bug fixes for database board text
This patch fixes extra '\0' recorded in text of database records, which are accidentally caused by the use of std::string::append(). It will automatically remove such '\0' in the string if there is any.

Also, when there are board texts recorded but not corresponding child database records, the board texts are not shown which may cause confusion. This patch fixes this, and all children with either a record or a board text will be displayed.

[skip_fishtest]