introduce multi visit to reduce NNUE cost
Instead of performing a single visit per playout, we conduct multiple visits per playout. This approach reduces the incremental update cost as we traverse down and up through the graph, leading to an increase in the number of visits per second. As a result, the original best-first MCTS search now incorporates some "depth-first" characteristics.

However, this method introduces potential bias into the search. To prevent the addition of too many nodes at once, which could distort the visit distribution among child nodes, we limit the number of new visits added to a node to a specified proportion of the parent node's visits.