propagate value bound in tree so we can get mate steps in root
Currently bound does not affect the backpropagation, selection in MCTS. Only when we select best move at the root node, we will downweight all other non-mate moves if we are sure that the root node is a winning node.

Also, when we only one legal move at root, no node recycling is performed.